+++ src/__init__.py
@@ -1,0 +1,11 @@
+# Import any necessary modules or packages here
+
+# Define any global variables or constants here
+
+# Define any helper functions or classes here
+
+# Define the main entry point of your module here
+
+if __name__ == "__main__":
+    # Code to be executed when the module is run as a standalone script
+    pass
