+++ src/document_loaders/CompleteDIrectoryLoader.py
@@ -1,8 +1,8 @@
-import DocumentLoader
+import document_loaders.DocumentLoaderBase as DocumentLoaderBase
 from langchain_community.document_loaders import DirectoryLoader
 
 
-class CompleteDirectoryLoader(DocumentLoader):
+class CompleteDirectoryLoader(DocumentLoaderBase):
     """
     Concrete implementation of a document loader for directories.
 
@@ -19,12 +19,22 @@ class CompleteDirectoryLoader(DocumentLoader):
         loader = CompleteDirectoryLoader('/path/to/directory')
         documents = loader.load()
     """
+
     def __init__(self, directory_path):
         self.directory_path = directory_path
 
     def load(self):
+        """
+        Loads the documents from the specified directory and returns them.
+
+        Returns:
+            list: A list of loaded documents.
+
+        Example:
+            loader = CompleteDirectoryLoader('/path/to/directory')
+            documents = loader.load()
+        """
         loader = DirectoryLoader(self.directory_path)
         docs = loader.load()
         print("Loaded directory")
         return docs
-
