+++ src/document_loaders/__init__.py
@@ -1,0 +1,6 @@
+from .CompleteDIrectoryLoader import CompleteDirectoryLoader
+from .CSVDocumentLoader import CSVDocumentLoader
+from .JSONDocumentLoader import JSONDocumentLoader
+from .MarkdownDocumentLoader import MarkdownDocumentLoader
+from .PDFDocumentLoader import PDFDocumentLoader
+from .WebDocumentLoader import WebDocumentLoader
