+++ src/text_processors/__init__.py
@@ -1,0 +1,2 @@
+from .RCTS import RCTS
+from .RSJSON import RSJSON
