+++ src/vector_stores/__init__.py
@@ -1,0 +1,1 @@
+from .FAISSVectorStore import FAISSVectorStore
